requirements/base.txt
Django==5.2.6
ruff==0.12.11
djangorestframework==3.16.1
pytest==8.4.1
pytest-django==4.11.1
django-model-utils==5.0.0